+++ packages/legacy/CcsdsFrameStripper.cpp
@@ -78,7 +78,7 @@ CommandableObject* CcsdsFrameStripper::createObject(CommandProcessor* cmd_proc, 
     uint8_t sync_marker[MAX_STR_SIZE];
     if(!StringLib::match(sync_str, "NONE"))
     {
-        sync_size = (int)StringLib::size(sync_str);
+        sync_size = StringLib::size(sync_str);
         if(sync_size <= 0 || sync_size % 2 != 0)
         {
             mlog(CRITICAL, "Sync marker is an invalid length: %d", sync_size);
